Task 5 fourth commit/added result

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -7,7 +7,7 @@ public class NumberDivider {
 
         toBeDivided = 19;
 
-        float result = toBeDivided / 5;
+        float result = (float) toBeDivided / 5;
 
         System.out.println(result);
